Seperate some Endpoints into Routes This is the start of the seperation. Will try to make the applicaiton more modular.
