+++ templates/base.html
@@ -8,6 +8,8 @@
     <script src="{{url_for('static', filename='script/jquery3.6.0.min.js')}}"></script>
     <script src="{{url_for('static', filename='DataTables/datatables.min.js')}}"></script>
     <script src="{{url_for('static', filename='script/script.js')}}"></script>
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.2/css/all.min.css" integrity="sha512-z3gLpd7yknf1YoNbCzqRKc4qyor8gaKU1qmn+CShxbuBusANI9QpRohGBreCFkKxLhei6S9CQXFEbbKuqLg0DA==" crossorigin="anonymous" referrerpolicy="no-referrer" />   
+    
 </head>
 <body>
     {% include 'header.html' %}
