+++ templates/catalog.html
@@ -1,7 +1,7 @@
 {% extends 'base.html' %}
 
 {% block content %}
-    <p>This is the main thing! Hope it works.</p>
+    
 {% endblock %}
