+++ templates/header.html
@@ -1,5 +1,5 @@
 <!-- Your navigation links, logo, and other header content -->
 <div class="header">
     <h1>PNGUMIT STORE</h1>
-    <h3 class="logout-button"><a href="{{ url_for('logout')}}">Logout</a></h3>
+    <h3 class="logout-button">{{session['userName']}} <a href="{{ url_for('logout')}}"><i class="fa fa-sign-out" aria-hidden="true"></i></a></h3>
 </div>
