+++ templates/invoice_preview.html
@@ -70,7 +70,7 @@
 
       <div class="controls">
         <!-- Cancel button -->
-        <button><a href="{{url_for('generate_pdf')}}">Confirm</a></button>
+        <button><a href="{{url_for('route_generatePDF.generate_pdf')}}">Confirm</a></button>
         <a href="{{url_for('cancel')}}">Cancel</a>
     </div>
