+++ templates/sidepanel.html
@@ -1,6 +1,6 @@
 <div class="left-panel">
     <a href="{{ url_for('home') }}"> <button>Home</button></a> 
-    <a href="{{ url_for('invoice_page') }}"><button>Invoice</button></a> 
+    <a href="{{ url_for('route_generatePDF.invoice_page') }}"><button>Invoice</button></a> 
     <a href="{{ url_for('inventory_page') }}"><button>Inventory</button></a>
     <a href="{{ url_for('catalog_page') }}"><button>Item Catalog</button></a>
     <a href="{{ url_for('customer_page') }}"><button>Customers</button></a>
